Add HVAC-schedule API-methods. (#94)
* Add API methods for programming a HVAC schedule

Apparently, in newer (PH2) Zoes, the HVAC schedule is now programmed
through a v2 schedules endpoint very much similar to how
charge-schedules are handled. This commit adds a very bsich backend
API implementation for this.

Note that there is some clarify missing around timestamp-handling and
possiblities. These are ignored in this simple implementation, but
assumed to be added later when more knowledge on the topic is gained.

* Log HTTP-method used on kamereon requests

* Add tests for hvac-settings and hvac-schedule endpoints

* Move chargemode for_json to action test-module

* Add simplistic hvac schedule action test